requirements.txt
face_recognition==1.3.0
numpy==2.1.3
opencv_python==4.10.0.84
tensorflow_cpu==2.18.0
